@@ -1,6 +1,6 @@
 {
   "name": "node-static",
-  "version": "0.7.11",
+  "version": "0.8.0",
   "description": "simple, compliant file streaming module for node",
   "author": "Alexis Sellier <alexis@cloudhead.io>",
   "contributors": [
@@ -33,15 +33,15 @@
   "scripts": {
     "build": "rollup -c",
     "start": "./bin/cli.js",
-    "lint": "eslint --ext=cjs,js .",
+    "lint": "eslint .",
     "mocha": "mocha test/integration --parallel",
     "test": "c8 npm run mocha"
   },
   "dependencies": {
     "colors": "1.4.0",
-    "command-line-basics": "^1.0.2",
+    "command-line-basics": "^2.0.1",
     "mime": "^3.0.0",
-    "minimatch": "^5.0.1"
+    "minimatch": "^9.0.0"
   },
   "c8": {
     "exclude": [
@@ -50,15 +50,17 @@
     ]
   },
   "devDependencies": {
-    "c8": "^7.11.0",
-    "chai": "^4.3.6",
-    "eslint": "^8.10.0",
-    "mocha": "^9.2.1",
+    "@eslint/js": "^9.9.0",
+    "c8": "^10.1.2",
+    "chai": "^5.1.1",
+    "eslint": "^9.9.0",
+    "globals": "^15.9.0",
+    "mocha": "^10.7.3",
     "node-fetch": "^2.6.6",
-    "rollup": "^2.70.0"
+    "rollup": "^4.21.0"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=18.0.0"
   },
   "bugs": {
     "url": "https://github.com/cloudhead/node-static/issues"
