@@ -5,7 +5,7 @@ import path from 'path';
 // import buffer from 'buffer';
 
 import mime from 'mime';
-import minimatch from 'minimatch';
+import {minimatch} from 'minimatch';
 import {mstat} from './node-static/util.js';
 
 const pkg = JSON.parse(fs.readFileSync(
@@ -208,7 +208,7 @@ class Server {
         try {
             pathname = decodeURIComponent(new URL(req.url, 'http://localhost').pathname);
         }
-        catch(e) {
+        catch {
             return process.nextTick(function() {
                 return finish(400, {});
             });
